* Adding timer for validation account in 30 minutes

+++ app/src/main/java/id/exomatik/mushafmuslim/model/ModelTransaction.kt
@@ -8,5 +8,5 @@ data class ModelTransaction(
     var idTransaction: String = "",
     var date: String = "",
     var username: String = "",
-    var poin: Long = 0
+    var poin: Int = 0
     ) : Parcelable
